+++ src/ArtifactsMMO.NET/Queries/CharacterAchievementsQuery.cs
@@ -56,7 +56,8 @@ namespace ArtifactsMMO.NET.Queries
             }
 
             var queryStringBuilder = new QueryStringBuilder();
-            queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(Type)), Type?.ToString().ToLower());
+            queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(Type)),
+                JsonNamingPolicy.SnakeCaseLower.ConvertName(Type?.ToString() ?? string.Empty));
             queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(Completed)), Completed?.ToString());
             queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(Page)), Page?.ToString());
             queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(Size)), Size?.ToString());