+++ src/ArtifactsMMO.NET/Queries/ItemsQuery.cs
@@ -90,9 +90,11 @@ namespace ArtifactsMMO.NET.Queries
 
             var queryStringBuilder = new QueryStringBuilder();
             queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(CraftMaterial)), CraftMaterial);
-            queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(CraftSkill)), CraftSkill?.ToString().ToLower());
+            queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(CraftSkill)), 
+                JsonNamingPolicy.SnakeCaseLower.ConvertName(CraftSkill?.ToString() ?? string.Empty));
             queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(Name)), Name);
-            queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(Type)), Type?.ToString().ToLower());
+            queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(Type)),
+                JsonNamingPolicy.SnakeCaseLower.ConvertName(Type?.ToString() ?? string.Empty));
             queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(MaxLevel)), MaxLevel?.ToString());
             queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(MinLevel)), MinLevel?.ToString());
             queryStringBuilder.AddParameter(JsonNamingPolicy.SnakeCaseLower.ConvertName(nameof(Page)), Page?.ToString());